fix/package-dts-error
